project: move public/exported includes to QtZeroConf/
This allows this project to be easily consumed either as an installed
package or via cmake's FetchContent, transparent to the consuming project.
As with many other projects, the header files are (already) installed to
a project subfolder, as a 'namespacing' technique to avoid collision with
other projects. This is good practice, but in order to also be used via
FetchContent the header files need to be in a consistent folder hierarchy.

This change moves the global/installed headers to a QtZeroConf/ subfolder,
and adjusts the sources which use it.

Signed-off-by: Marc Reilly <marc@cpdesign.com.au>